pyproject.toml
[tool.poetry]
name = "airogue"
version = "0.1.0"
description = ""
authors = ["Tarolrr"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.14.2"
tcod = "^16.2.2"
tcod-ecs = "^5.1.0"
langchain-core = "^0.1.33"
langchain-community = "^0.0.29"
langchain-openai = "^0.1.0"
langchain = "^0.1.13"
pyyaml = "^6.0.1"
colored = "^2.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
